feat: `OpenAi` now returns `UsageData` for `generateResponse` and `generateStreamingResponse` methods.
BREAKING CHANGE: converted `OpenAi` static methods to be instance methods. Also updated params to be a single object for most methods in `OpenAi` and `Conversation`.

Made these changes to simplify the maintenance of `OpenAi` and also to simplify the code calling the api to not need order args (and pass in undefined often) with such a large number of optional parameters.